field changed to decimal

+++ Raffle/Models/Raffle.cs
@@ -3,7 +3,7 @@
     public class Raffle
     {
         public int Id { get; set; }
-        public int prizeValue {  get; set; } 
+        public decimal prizeValue {  get; set; } 
         public  DateTime dataRaffle {  get; set; }
         public string result { get; set; }
         public DateTime CreatedAt { get; set; }
